Convert pdf 2 set of images and save
Use pdf2img library to convert uplaoded .pdf to a list of image and save
each of them as a DeckImage belonging to a pitch deck.
This currently works but is somewhat slow and there is no indicator on
the user side.
There is also a basic error check which should probably be moved to
form validation override if time allows.